% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_delete.R
\name{hs_delete}
\alias{hs_delete}
\title{Delete an annotation}
\source{
\url{https://h.readthedocs.io/en/latest/api/#delete}
}
\usage{
hs_delete(token, id)
}
\arguments{
\item{token}{Character. Your account token, which you can generate at \url{https://hypothes.is/register}}

\item{id}{Character. A hypothes.is annotation id.}
}
\value{
TRUE on successful deletion.
}
\description{
Delete an annotation
}

