% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_read.R
\name{hs_read}
\alias{hs_read}
\title{Retrieve a single annotation by ID}
\source{
\url{https://h.readthedocs.io/en/latest/api/#read}
}
\usage{
hs_read(id)
}
\arguments{
\item{id}{Character. A hypothes.is annotation id.}
}
\value{
A dataframe with the contents of that annotation.
}
\description{
Retrieve a single annotation by ID
}
\examples{
hs_read("WFMnSC3FEeaNvLeGkQeJbg")
}

