% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_utils.R
\name{hs_reply}
\alias{hs_reply}
\title{Create a reply to a comment}
\usage{
hs_reply(token, user, id, text, ...)
}
\arguments{
\item{token}{Character. Your account token, which you can generate at \url{https://hypothes.is/register}}

\item{user}{Character. Your user account, normally in the format \code{acct:username@hypothes.is}}

\item{id}{Character. The annotation ID to reply to.}

\item{text}{Character. Text to put in the body of the annotation. This will be coerced into a character vector of length 1 using \link{paste}.}

\item{...}{Other arguments to pass to \link{hs_create}.}
}
\description{
This is a utility wrapper around \link{hs_create} that takes an original comment ID and creates a reply to it by adding the custom \code{references} field when constructing the annotation. Normal fields like
}

