% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisr-package.R
\docType{package}
\name{hypothesisr}
\alias{hypothesisr}
\alias{hypothesisr-package}
\title{hypothesisr: Wrapper for the Hypothes.is API}
\description{
Interact with the API for the web annotation service hypothes.is.
Allows users to add, search for, and retrieve annotation data.
}

