% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_utils.R
\name{open_}
\alias{open_}
\alias{open_annotation}
\alias{open_context}
\alias{open_uri}
\title{Functions for opening an annotation in the system browser}
\usage{
open_context(id)

open_annotation(id)

open_uri(id)
}
\arguments{
\item{id}{Annotation ID.}
}
\description{
Functions for opening an annotation in the system browser
}
\section{Functions}{
\itemize{
\item \code{open_context}: Open an annotation in context, displaying the original webpage with an annotation overlay

\item \code{open_annotation}: Open an annotation on the hypothes.is webpage

\item \code{open_uri}: Open the original webpage to which the annotation links
}}

